Upgrade to GeoMapFish master
