+++ .gitignore
@@ -7,6 +7,7 @@
 /apache/geoportailv3.wsgi
 /deploy/deploy.cfg
 /alembic.ini
+/alembic_static.ini
 /lux_alembic.ini
 /development.ini
 /production.ini
