+++ .travis.yml
@@ -4,7 +4,7 @@ python:
 - 2.7
 
 install:
-- sudo apt-get install postgis 
+- sudo apt-get install postgis apache2
 - sudo -u postgres createdb -E UTF8 -T template0 template_postgis
 - sudo -u postgres psql -d template_postgis -c "CREATE EXTENSION postgis;"
 - sudo -u postgres createdb -E UTF8 -T template_postgis lux
@@ -17,12 +17,12 @@ install:
 - sudo -u postgres psql -d lux -c 'GRANT ALL ON SCHEMA geov3_static TO "www-data";'
 - sudo -u postgres psql -d lux -c 'GRANT ALL ON geometry_columns TO "www-data";'
 
-- sudo mkdir -p /srv/tomcat/tomcat1/webapps
-- sudo chmod g+w,o+w /srv/tomcat/tomcat1/webapps
+- sudo mkdir -p /srv/tomcat/tomcat1/webapps /var/www/vhosts/luxembourg-geomapfish/conf
+- sudo chmod g+w,o+w /srv/tomcat/tomcat1/webapps /var/www/vhosts/luxembourg-geomapfish/conf
 
 - make -f travis.mk .build/requirements.timestamp alembic.ini lux_alembic.ini
-- .build/venv/bin/alembic upgrade head 
-- .build/venv/bin/alembic -c lux_alembic.ini upgrade head 
+- .build/venv/bin/alembic upgrade head
+- .build/venv/bin/alembic -c lux_alembic.ini upgrade head
 - make -f travis.mk build
 
 script:
