+++ CONST_alembic/main/versions/164ac0819a61_add_image_format_to_wmts_layer.py
@@ -1,0 +1,54 @@
+# -*- coding: utf-8 -*-
+
+# Copyright (c) 2015, Camptocamp SA
+# All rights reserved.
+
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are met:
+
+# 1. Redistributions of source code must retain the above copyright notice, this
+#    list of conditions and the following disclaimer.
+# 2. Redistributions in binary form must reproduce the above copyright notice,
+#    this list of conditions and the following disclaimer in the documentation
+#    and/or other materials provided with the distribution.
+
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+# ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+# DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+# ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+# ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+# The views and conclusions contained in the software and documentation are those
+# of the authors and should not be interpreted as representing official policies,
+# either expressed or implied, of the FreeBSD Project.
+
+"""Add image format to WMTS layer
+
+Revision ID: 164ac0819a61
+Revises: 20137477bd02
+Create Date: 2015-03-06 09:08:05.754746
+"""
+
+from alembic import op, context
+from sqlalchemy import Column, Unicode
+
+# revision identifiers, used by Alembic.
+revision = "164ac0819a61"
+down_revision = "20137477bd02"
+
+
+def upgrade():
+    schema = context.get_context().config.get_main_option("schema")
+
+    op.add_column("layer_wmts", Column("image_type", Unicode(10)), schema=schema)
+
+
+def downgrade():
+    schema = context.get_context().config.get_main_option("schema")
+
+    op.drop_column("layer_wmts", "image_type", schema=schema)
