--- CONST_alembic/versions/3f89a7d71a5e_alter_column_url_to_remove_limitation.py
@@ -1,51 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Copyright (c) 2013-2014, Camptocamp SA
-# All rights reserved.
-
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions are met:
-
-# 1. Redistributions of source code must retain the above copyright notice, this
-#    list of conditions and the following disclaimer.
-# 2. Redistributions in binary form must reproduce the above copyright notice,
-#    this list of conditions and the following disclaimer in the documentation
-#    and/or other materials provided with the distribution.
-
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-# ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-# DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
-# ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-# ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
-# The views and conclusions contained in the software and documentation are those
-# of the authors and should not be interpreted as representing official policies,
-# either expressed or implied, of the FreeBSD Project.
-
-
-"""Alter_column_url_to_remove_limitation
-
-Revision ID: 3f89a7d71a5e
-Revises: 20137477bd02
-Create Date: 2014-12-18 10:27:52.263992
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '3f89a7d71a5e'
-down_revision = '20137477bd02'
-
-from alembic import op, context
-from sqlalchemy import types
-
-def upgrade():
-    schema='%s_static' % context.get_context().config.get_main_option('schema')
-    op.alter_column('shorturl', 'url', type_=types.Unicode, schema=schema)
-
-def downgrade():
-    pass
