+++ CONST_vars.yaml
@@ -25,7 +25,7 @@ vars:
     # database parent schema
     parentschema: ""
     # apache/mapserver.conf "Allow" statement
-    mapserv_allow: Allow from 127.0.0.1
+    mapserv_allow: Allow from 127.0.0.1 ::1
     # mapserver connexion string
     mapserver_connection: user={dbuser} password={dbpassword} dbname={db} host={dbhost}
     # mapserver join tables
@@ -40,24 +40,14 @@ vars:
         '"default_role_id" "-1"
         "role_id" "^-?[0-9]*$"'
     # mapfile data subselect
-    mapfile_data_subselect: 'SELECT ST_Collect(ra.area) FROM main.restrictionarea AS ra, main.role_restrictionarea AS rra, main.layer_restrictionarea AS lra, main.treeitem AS la WHERE rra.role_id: %role_id% AND rra.restrictionarea_id: ra.id AND lra.restrictionarea_id: ra.id AND lra.layer_id: la.id AND la.name:'
+    mapfile_data_subselect: 'SELECT ST_Collect(ra.area) FROM main.restrictionarea AS ra, main.role_restrictionarea AS rra, main.layer_restrictionarea AS lra, main.treeitem AS la WHERE rra.role_id = %role_id% AND rra.restrictionarea_id = ra.id AND lra.restrictionarea_id = ra.id AND lra.layer_id = la.id AND la.name = '
     # mapserver filter on role id only
-    mapfile_data_noarea_subselect: 'SELECT rra.role_id FROM {mapserver_join_tables} WHERE rra.restrictionarea_id: ra.id AND lra.restrictionarea_id: ra.id AND lra.layer_id: la.id AND la.name:'
-    # print path
-    print_path: /print-c2cgeoportal-{instanceid}/pdf/
+    mapfile_data_noarea_subselect: 'SELECT rra.role_id FROM {mapserver_join_tables} WHERE rra.restrictionarea_id = ra.id AND lra.restrictionarea_id = ra.id AND lra.layer_id = la.id AND la.name = '
     # port for the waitress web server
     waitress_port: 5000
-    # password change replication activation
-    auth_replication_enabled: false
-    # database connection for auth replication
-    dbhost_replication: to_be_defined
-    dbport_replication: to_be_defined
-    db_replication: to_be_defined
-    dbuser_replication: to_be_defined
-    dbpassword_replication: to_be_defined
     enable_admin_interface: True
 
-    directory: pwd
+    directory: pwd | tr -d '\n'
     python_path: __import__('distutils').sysconfig.get_python_lib()
 
     # Authentication settings
@@ -68,8 +58,6 @@ vars:
     # Database informations
     sqlalchemy:
         url: postgresql://{dbuser}:{dbpassword}@{dbhost}:{dbport}/{db}
-    sqlalchemy_replication:
-        url: postgresql://{dbuser_replication}:{dbpassword_replication}@{dbhost_replication}:{dbport_replication}/{db_replication}
 
     # 10 days of default cache
     default_max_age: 864000
@@ -79,8 +67,16 @@ vars:
         config: "{directory}/jsbuild/app.cfg"
         root_dir: "{directory}"
 
-    # pyramid_closure configuration
+    # used for the "node_modules" and "closure" static views
     closure_library_path: 'process.stdout.write(require("openlayers/node_modules/closure-util").getLibraryPath())'
+    node_modules_path: "{directory}/node_modules"
+
+    # pyramid_closure configuration
+    # Each item in the roots_with_prefix array is an array with two elements. The
+    # first element is the path pyramid_closure passed to request.static_url. The
+    # second element is the file system path where js source files are searched.
+    # For this to work window.CLOSURE_BASE_PATH being set to the empty string in
+    # the HTML page.
     pyramid_closure:
         roots_with_prefix:
         - ["{closure_library_path}/closure/goog", "{closure_library_path}/closure/goog"]
@@ -88,8 +84,7 @@ vars:
         - ["{node_modules_path}/openlayers", "{node_modules_path}/openlayers"]
         - ["{node_modules_path}/ngeo/src", "{node_modules_path}/ngeo/src"]
 
-    # used for the "node_modules" and "closure" static views
-    node_modules_path: "{directory}/node_modules"
+    layers: {}
 
     # The application's default language. This is the language used by
     # the application if no specific language is specified in the URLs.
@@ -166,10 +161,18 @@ vars:
     # this application is a child application in a parent/child
     # architecture.
     external_themes_url:
-    external_mapserv_url:
-    # If the WFS service is different from the WMS service, uncomment and set the
-    # external_mapserv_wfs_url parameter
-    #external_mapserv_wfs_url:
+
+    mapserverproxy:
+        # URL to internal MapServer instance
+        mapserv_url: http://localhost/{instanceid}/mapserv
+        # If the WFS service is different from the WMS service, uncomment and set the
+        # mapserv_wfs_url parameter
+        #mapserv_wfs_url:
+
+        external_mapserv_url:
+        # If the WFS service is different from the WMS service, uncomment and set the
+        # external_mapserv_wfs_url parameter
+        #external_mapserv_wfs_url:
 
     servers:
 
@@ -179,14 +182,8 @@ vars:
 
     # The shortner base configuration
     shortener:
-      # The base of created URL
-      base_url:  http://{host}{apache_entry_point}s/
-
-    # URL to internal MapServer instance
-    mapserv_url: http://localhost/{instanceid}/mapserv
-    # If the WFS service is different from the WMS service, uncomment and set the
-    # mapserv_wfs_url parameter
-    #mapserv_wfs_url:
+        # The base of created URL
+        base_url:  http://{host}{apache_entry_point}s/
 
     # Define whether the MapServer proxy should hide the OGC capabilities.
     hide_capabilities: false
@@ -199,7 +196,7 @@ vars:
     use_security_metadata: false
 
     # For print proxy
-    print_url: http://localhost:8080/print-c2cgeoportal-{instanceid}/pdf/
+    print_url: http://localhost:8080/print-{instanceid}/print/{package}
 
     # For base layers
     tiles_url:
@@ -217,12 +214,6 @@ vars:
     #proxies:
     #    http: http://someproxy
 
-    wfs:
-        # Set to True so that MapserverProxy limits the number of features
-        # returned by a WFS GetFeature request:
-        enable_limit_featurecollection: True
-        # Max number of features returned by a WFS GetFeature request:
-        maxfeatures: 200
 
     # Checker configuration
     checker:
@@ -231,6 +222,7 @@ vars:
         print_center_lat: to be defined
         print_scale: 10000
         fulltextsearch: text to search
+        lang_files: [ngeo, cgxp, cgxp-api]
 
     # Check collector configuration
     check_collector:
@@ -245,10 +237,14 @@ vars:
               display: Edit page
             - name: checker_edit_js
               display: Edit script
+            - name: checker_theme_errors
+              display: Errors in the themes
             - name: checker_api
               display: API
             - name: checker_xapi
               display: Extended API
+            - name: checker_lang_files
+              display: Language files
             - name: checker_printcapabilities
               display: Print capabilities
             - name: checker_pdf
@@ -287,5 +283,3 @@ interpreted:
     - development
     bash:
     - directory
-    yaml:
-    - directory # to trim
